Reuploading with debug onclick events
