--- Scripts/KeyboardMotion.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0fa4d02eaad6ace40873146660a908b6
-timeCreated: 1505333061
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
